+++ src/Hacks/AllModesPlatformer.cpp
@@ -5,8 +5,6 @@
 
 using namespace geode::prelude;
 
-#ifndef GEODE_IS_WINDOWS
-
 class $modify (GJBaseGameLayer)
 {
     void collisionCheckObjects(PlayerObject* p0, gd::vector<GameObject*>* p1, int p2, float p3)
@@ -22,7 +20,17 @@ class $modify (GJBaseGameLayer)
                     if(this->canBeActivatedByPlayer(p0, as<EffectGameObject*>(obj)))
                     {
                         this->playerWillSwitchMode(p0, obj);
-                        this->switchToFlyMode(p0, obj, false, as<int>(obj->m_objectType));                        
+                        #ifdef GEODE_IS_WINDOWS
+                        p0->switchedToMode(obj->m_objectType);
+
+                        if (obj->m_objectType == GameObjectType::SwingPortal)
+                            p0->toggleSwingMode(true, false);
+                        else
+                            p0->toggleDartMode(true, false);
+                        
+                        #else
+                        this->switchToFlyMode(p0, obj, false, as<int>(obj->m_objectType));
+                        #endif
                         obj->playShineEffect();
                     }
                 }
@@ -34,5 +42,3 @@ class $modify (GJBaseGameLayer)
 
     QOLMOD_MOD_HOOK("all-plat", "GJBaseGameLayer::collisionCheckObjects")
 };
-
-#endif